rename component related to countries page
